backend/requirements.txt
# FastAPI and server
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Data validation and settings
pydantic>=2.6.1
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Database (MongoDB with Beanie ODM)
motor>=3.3.2
pymongo>=4.6.1
beanie>=1.24.0

# Authentication and security
bcrypt>=4.1.2
PyJWT>=2.8.0

# File processing
PyPDF2>=3.0.1
python-docx>=1.1.0
pdfminer.six>=20231228

# AI/ML services
openai>=1.3.7
google-generativeai>=0.5.4

# Development and testing (optional)
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2
